#28: Changed all conception of values and params. (#29)
Need to support lists of parameters in queries, so abstraction should be on values level but not in parameter(name + value) level. Changed all conception and interfaces and big part of architecture. Now we need to detect the type of parameter in values. Also removed all factories and works with Class detections = more clear OOP.

Co-authored-by: asysoev <asysoev@argenta.dev>